CI: Support cached files.
TravisCI and SemaphoreCI support caching to speed up builds
by avoiding having to re-download and re-build binaries and
libraries built successfully in the last build.

Update the CI setup script to only perform the actions it
needs to.

+++ .ci/ci-setup.sh
@@ -142,31 +142,81 @@ function compile {
 	popd
 }
 
+# Determined if the specified command has already been installed
+function cmd_installed {
+    local cmd="$1"
+
+    local path="${prefix_dir}/bin/$cmd"
+
+    [ -e "$path" ]
+}
+
+# Determine if the specified library version is available
+function lib_installed {
+    local name="$1"
+    local required_version="$2"
+
+    version=$(pkg-config --print-provides "$name" 2>/dev/null | awk '{print $3}')
+
+    [ "$version" = "$required_version" ]
+}
+
 pushd "$deps_dir"
 
 # Build glib
 glib_major=`echo $glib_version | cut -d. -f1`
 glib_minor=`echo $glib_version | cut -d. -f2`
-curl -L -O "$gnome_dl/glib/${glib_major}.${glib_minor}/glib-${glib_version}.tar.xz"
-compile glib glib-${glib_version}.tar.xz glib-${glib_version}
+file="glib-${glib_version}.tar.xz"
+
+if ! lib_installed "glib-2.0" "$glib_version"
+then
+    if [ ! -e "$file" ]
+    then
+        curl -L -O "$gnome_dl/glib/${glib_major}.${glib_minor}/$file"
+    fi
+
+    compile glib glib-${glib_version}.tar.xz glib-${glib_version}
+fi
 
 # Build json-glib
 json_major=`echo $json_glib_version | cut -d. -f1`
 json_minor=`echo $json_glib_version | cut -d. -f2`
-curl -L -O "$gnome_dl/json-glib/${json_major}.${json_minor}/json-glib-${json_glib_version}.tar.xz"
-compile json-glib json-glib-${json_glib_version}.tar.xz json-glib-${json_glib_version}
+file="json-glib-${json_glib_version}.tar.xz"
+
+if ! lib_installed "json-glib-1.0" "$json_glib_version"
+then
+    if [ ! -e "$file" ]
+    then
+        curl -L -O "$gnome_dl/json-glib/${json_major}.${json_minor}/$file"
+    fi
+
+    compile json-glib json-glib-${json_glib_version}.tar.xz json-glib-${json_glib_version}
+fi
 
 # Build check
 # We need to build check as the check version in the OS used by travis isn't
 # -pedantic safe.
-curl -L -O "https://github.com/libcheck/check/releases/download/${check_version}/check-${check_version}.tar.gz"
-compile check check-${check_version}.tar.gz check-${check_version}
+if ! lib_installed "check" "${check_version}"
+then
+    file="check-${check_version}.tar.gz"
 
-# Install bats
-git clone https://github.com/sstephenson/bats.git
-pushd bats
-sudo ./install.sh "$prefix_dir"
-popd
+    if [ ! -e "$file" ]
+    then
+        curl -L -O "https://github.com/libcheck/check/releases/download/${check_version}/$file"
+    fi
+
+    compile check check-${check_version}.tar.gz check-${check_version}
+fi
+
+cmd="bats"
+if ! cmd_installed "$cmd"
+then
+    # Install bats
+    [ ! -d bats ] && git clone https://github.com/sstephenson/bats.git
+    pushd bats
+    sudo ./install.sh "$prefix_dir"
+    popd
+fi
 
 if [ "$nested" != "Y" ]
 then
@@ -174,83 +224,98 @@ then
     exit 0
 fi
 
-# build gcc (required for qemu-lite)
-gcc_dir="gcc-${gcc_version}"
-gcc_site="http://mirrors.kernel.org/gnu/gcc/${gcc_dir}"
-gcc_file="gcc-${gcc_version}.tar.bz2"
-gcc_url="${gcc_site}/${gcc_file}"
-curl -L -O "$gcc_url"
+cmd="gcc"
+if ! cmd_installed "$cmd"
+then
+    # build gcc (required for qemu-lite)
+    gcc_dir="gcc-${gcc_version}"
+    gcc_site="http://mirrors.kernel.org/gnu/gcc/${gcc_dir}"
+    gcc_file="gcc-${gcc_version}.tar.bz2"
+    gcc_url="${gcc_site}/${gcc_file}"
 
-gcc_opts=""
-gcc_opts+=" --enable-languages=c"
-gcc_opts+=" --disable-multilib"
-gcc_opts+=" --disable-libstdcxx"
-gcc_opts+=" --disable-bootstrap"
-gcc_opts+=" --disable-nls"
-gcc_opts+=" --prefix=\"${prefix_dir}\""
+    if [ ! -e "$gcc_file" ]
+    then
+        curl -L -O "$gcc_url"
+    fi
 
-compile gcc "$gcc_file" "$gcc_dir" "$gcc_opts"
+    gcc_opts=""
+    gcc_opts+=" --enable-languages=c"
+    gcc_opts+=" --disable-multilib"
+    gcc_opts+=" --disable-libstdcxx"
+    gcc_opts+=" --disable-bootstrap"
+    gcc_opts+=" --disable-nls"
+    gcc_opts+=" --prefix=\"${prefix_dir}\""
+
+    compile gcc "$gcc_file" "$gcc_dir" "$gcc_opts"
+fi
 
 # Use built version of gcc
 export CC="${prefix_dir}/bin/gcc"
 
 # build qemu-lite
-qemu_lite_site="https://github.com/01org/qemu-lite/archive/"
-qemu_lite_file="${qemu_lite_version}.tar.gz"
-qemu_lite_url="${qemu_lite_site}/${qemu_lite_file}"
-qemu_lite_dir="qemu-lite-${qemu_lite_version}"
+cmd="qemu-system-x86_64"
+if ! cmd_installed "$cmd"
+then
+    qemu_lite_site="https://github.com/01org/qemu-lite/archive/"
+    qemu_lite_file="${qemu_lite_version}.tar.gz"
+    qemu_lite_url="${qemu_lite_site}/${qemu_lite_file}"
+    qemu_lite_dir="qemu-lite-${qemu_lite_version}"
 
-qemu_lite_opts=""
-qemu_lite_opts+=" --disable-bluez"
-qemu_lite_opts+=" --disable-brlapi"
-qemu_lite_opts+=" --disable-bzip2"
-qemu_lite_opts+=" --disable-curl"
-qemu_lite_opts+=" --disable-curses"
-qemu_lite_opts+=" --disable-debug-tcg"
-qemu_lite_opts+=" --disable-fdt"
-qemu_lite_opts+=" --disable-glusterfs"
-qemu_lite_opts+=" --disable-gtk"
-qemu_lite_opts+=" --disable-libiscsi"
-qemu_lite_opts+=" --disable-libnfs"
-qemu_lite_opts+=" --disable-libssh2"
-qemu_lite_opts+=" --disable-libusb"
-qemu_lite_opts+=" --disable-linux-aio"
-qemu_lite_opts+=" --disable-lzo"
-qemu_lite_opts+=" --disable-opengl"
-qemu_lite_opts+=" --disable-qom-cast-debug"
-qemu_lite_opts+=" --disable-rbd"
-qemu_lite_opts+=" --disable-rdma"
-qemu_lite_opts+=" --disable-sdl"
-qemu_lite_opts+=" --disable-seccomp"
-qemu_lite_opts+=" --disable-slirp"
-qemu_lite_opts+=" --disable-snappy"
-qemu_lite_opts+=" --disable-spice"
-qemu_lite_opts+=" --disable-strip"
-qemu_lite_opts+=" --disable-tcg-interpreter"
-qemu_lite_opts+=" --disable-tcmalloc"
-qemu_lite_opts+=" --disable-tools"
-qemu_lite_opts+=" --disable-tpm"
-qemu_lite_opts+=" --disable-usb-redir"
-qemu_lite_opts+=" --disable-uuid"
-qemu_lite_opts+=" --disable-vnc"
-qemu_lite_opts+=" --disable-vnc-{jpeg,png,sasl}"
-qemu_lite_opts+=" --disable-vte"
-qemu_lite_opts+=" --disable-xen"
-qemu_lite_opts+=" --enable-attr"
-qemu_lite_opts+=" --enable-cap-ng"
-qemu_lite_opts+=" --enable-kvm"
-qemu_lite_opts+=" --enable-virtfs"
-qemu_lite_opts+=" --target-list=x86_64-softmmu"
-qemu_lite_opts+=" --extra-cflags=\"-fno-semantic-interposition -O3 -falign-functions=32\""
-qemu_lite_opts+=" --prefix=\"${prefix_dir}\""
-qemu_lite_opts+=" --datadir=\"${prefix_dir}/share/qemu-lite\""
-qemu_lite_opts+=" --libdir=\"${prefix_dir}/lib64/qemu-lite\""
-qemu_lite_opts+=" --libexecdir=\"${prefix_dir}/libexec/qemu-lite\""
+    qemu_lite_opts=""
+    qemu_lite_opts+=" --disable-bluez"
+    qemu_lite_opts+=" --disable-brlapi"
+    qemu_lite_opts+=" --disable-bzip2"
+    qemu_lite_opts+=" --disable-curl"
+    qemu_lite_opts+=" --disable-curses"
+    qemu_lite_opts+=" --disable-debug-tcg"
+    qemu_lite_opts+=" --disable-fdt"
+    qemu_lite_opts+=" --disable-glusterfs"
+    qemu_lite_opts+=" --disable-gtk"
+    qemu_lite_opts+=" --disable-libiscsi"
+    qemu_lite_opts+=" --disable-libnfs"
+    qemu_lite_opts+=" --disable-libssh2"
+    qemu_lite_opts+=" --disable-libusb"
+    qemu_lite_opts+=" --disable-linux-aio"
+    qemu_lite_opts+=" --disable-lzo"
+    qemu_lite_opts+=" --disable-opengl"
+    qemu_lite_opts+=" --disable-qom-cast-debug"
+    qemu_lite_opts+=" --disable-rbd"
+    qemu_lite_opts+=" --disable-rdma"
+    qemu_lite_opts+=" --disable-sdl"
+    qemu_lite_opts+=" --disable-seccomp"
+    qemu_lite_opts+=" --disable-slirp"
+    qemu_lite_opts+=" --disable-snappy"
+    qemu_lite_opts+=" --disable-spice"
+    qemu_lite_opts+=" --disable-strip"
+    qemu_lite_opts+=" --disable-tcg-interpreter"
+    qemu_lite_opts+=" --disable-tcmalloc"
+    qemu_lite_opts+=" --disable-tools"
+    qemu_lite_opts+=" --disable-tpm"
+    qemu_lite_opts+=" --disable-usb-redir"
+    qemu_lite_opts+=" --disable-uuid"
+    qemu_lite_opts+=" --disable-vnc"
+    qemu_lite_opts+=" --disable-vnc-{jpeg,png,sasl}"
+    qemu_lite_opts+=" --disable-vte"
+    qemu_lite_opts+=" --disable-xen"
+    qemu_lite_opts+=" --enable-attr"
+    qemu_lite_opts+=" --enable-cap-ng"
+    qemu_lite_opts+=" --enable-kvm"
+    qemu_lite_opts+=" --enable-virtfs"
+    qemu_lite_opts+=" --target-list=x86_64-softmmu"
+    qemu_lite_opts+=" --extra-cflags=\"-fno-semantic-interposition -O3 -falign-functions=32\""
+    qemu_lite_opts+=" --prefix=\"${prefix_dir}\""
+    qemu_lite_opts+=" --datadir=\"${prefix_dir}/share/qemu-lite\""
+    qemu_lite_opts+=" --libdir=\"${prefix_dir}/lib64/qemu-lite\""
+    qemu_lite_opts+=" --libexecdir=\"${prefix_dir}/libexec/qemu-lite\""
 
-curl -L -O "${qemu_lite_url}"
+    if [ ! -e "$qemu_lite_file" ]
+    then
+        curl -L -O "${qemu_lite_url}"
+    fi
 
-compile qemu-lite "$qemu_lite_file" \
-"$qemu_lite_dir" "$qemu_lite_opts"
+    compile qemu-lite "$qemu_lite_file" \
+        "$qemu_lite_dir" "$qemu_lite_opts"
+fi
 
 # install kernel + Clear Containers image
 mkdir -p assets
@@ -269,7 +334,10 @@ clr_kernel=$(curl -l -s -L "${clr_kernel_base_url}" |\
 # download kernel
 if [ ! -e "${clr_assets_dir}/${clr_kernel}" ]
 then
-    curl -L -O "${clr_kernel_base_url}/${clr_kernel}"
+    if [ ! -e "$clr_kernel" ]
+    then
+        curl -L -O "${clr_kernel_base_url}/${clr_kernel}"
+    fi
 
     # install kernel
     # (note: cpio on trusty does not support "-D")
@@ -287,7 +355,7 @@ if [ ! -e "${clr_assets_dir}/${clr_image}" ]
 then
     for file in "${clr_image_url}-SHA512SUMS" "${clr_image_url}"
     do
-        curl -L -O "$file"
+        [ ! -e "$file" ] && curl -L -O "$file"
     done
 
     # verify image
